refactor: 전 단계(step1) 리뷰내용 반영
전체적으로 패키지 위치 수정 (오타로 인해 생긴 domain.domain 중복 패키지명 제거)

Numbers 클래스 리팩토링
- 배열을 List로 교체
- List를 조작하는 로직에 Stream 활용

StringAddCalculatorTest 클래스 리팩토링
- @ParameterizedTest 활용해 중복소스 제거
