+++ src/test/java/edu/nextstep/calculator/domain/NumbersTest.java
@@ -1,12 +1,11 @@
-package edu.nextstep.calculator.domain.domain;
+package edu.nextstep.calculator.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
+import java.util.Arrays;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import edu.nextstep.calculator.domain.Number;
-import edu.nextstep.calculator.domain.Numbers;
 
 @DisplayName("Numbers: Number 클래스 목록에 대한 일급 콜렉션")
 class NumbersTest {
@@ -19,7 +18,8 @@ class NumbersTest {
 			.isNotNull()
 			.isInstanceOf(Numbers.class)
 			.hasFieldOrPropertyWithValue("numbers",
-				new Number[] {Number.createFromString("1"), Number.createFromString("2"), Number.createFromString("3")});
+				Arrays.asList(Number.createFromString("1"), Number.createFromString("2"),
+					Number.createFromString("3")));
 	}
 
 	@DisplayName("sumAll: 인스턴스 변수 numbers 내 모든 Number 객체를 더한 후 원시값을 리턴함.")
